requirements.txt
alembic==1.3.2
attrs==19.3.0
Click==7.0
dotobject==1.2.1
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.12
wcwidth==0.1.8
Werkzeug==0.16.0
